separate email module
