+++ src/admin/admin.module.ts
@@ -5,12 +5,11 @@ import { Registrant } from '../entities/registrant.entity';
 import { AdminController } from './admin.controller';
 import { AdminLoginController } from './login.controller';
 import { AdminService } from './admin.service';
-import { EmailService } from './email.service';
+import { EmailModule } from 'src/email/email.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Registrant]), AuthModule],
+  imports: [TypeOrmModule.forFeature([Registrant]), AuthModule, EmailModule],
   controllers: [AdminController, AdminLoginController],
-  providers: [AdminService, EmailService],
-  exports: [EmailService]
+  providers: [AdminService],
 })
 export class AdminModule {}
